requirements.txt
google-api-python-client
python-digitalocean
ansible
